requirements.txt
estool==1.0
numpy==1.15.3
pandas==0.23.4
Pillow==5.3.0
python-dateutil==2.7.4
pytz==2018.6
six==1.11.0
torch==0.4.1
torchvision==0.2.1
